+++ src/index.ts
@@ -1,5 +1,9 @@
-export { default as build } from './actions/build';
+export { default as buildProjecg } from './actions/build-project';
+export { default as createProject } from './actions/create-project';
 export { default as exportProject } from './actions/export-project';
+export { default as initProject } from './actions/init-project';
+export { default as runMacro } from './actions/run-macro';
+export { default as testProject } from './actions/test-project';
 export { addins } from './addin';
 export { loadConfig } from './config';
 export { default as env } from './env';
